rgw_sal_motr: [CORTX-28987] Fix object deletion with null version
Problem : If version id parameter is passed as null to delete-object
API in case of unversioned bucket for deletion of simple object,
then object deletion does not happen.

Implementation : If instance is set to "null" we clear this string
in delete_object implementation. This implementation is similar to
what is done in rados.

Signed-off-by: Deepak Mahale <deepak.mahale@seagate.com>